backend/requirements.txt
fastapi
uvicorn[standard] # includes watchgod and python-dotenv
torch
torchvision
torchaudio
numpy
cython
## detectron2
pillow
python-multipart
python-dotenv
requests
openai
sqlite3

# installation for bert
transformers
sentence-transformers
pinecone
